% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtimeOutput.R, R/renderRealtime.R, R/shiny.R
\name{realtimeOutput}
\alias{realtimeOutput}
\alias{renderRealtime}
\alias{realtimeshiny}
\title{Wrapper functions for \code{realtime} in \code{shiny}}
\usage{
realtimeOutput(outputId, width = "100\%", height = 400)

renderRealtime(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{\code{character }output variable to read from.}

\item{width}{\code{numeric} width of the map (see \code{shinyWidgetOutput}).}

\item{height}{\code{numeric} height of the map (see
\code{shinyWidgetOutput}).}

\item{expr}{An expression that generates an HTML widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})?
This is useful if you want to save an expression in a variable.}
}
\description{
Use \code{leafletOutput()} to create a UI element, and
\code{renderLeaflet()} to render the widget.
}
