% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/realtime.R
\docType{package}
\name{realtime}
\alias{realtime}
\alias{realtime-package}
\alias{realtime}
\title{realtime}
\usage{
realtime(x, wait = 5, padding = 0, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{x}{\code{function} for fetching and process data.}

\item{wait}{\code{numeric} number of seconds to wait between plot
update.}

\item{padding}{\code{numeric} the padding of the plot.}

\item{width}{\code{numeric} the width of the plot.}

\item{height}{\code{numeric} the height of the plot.}
}
\description{
\emph{realtime} is an R package for creating data visualizations.

Create a live plot.
}
\details{
The argument to \code{x} should be a function that returns
  a \code{numeric} vector containing a \code{X} and \code{Y} coordinates
  for plotting in real-time.
}
\examples{
\donttest{
realtime(function() rnorm(1), 10)
}

}
